requirements.txt
flask
python-telegram-bot==13.15
docxtpl
python-docx
pydub
openai
pyTelegramBotAPI
python-dotenv
